Extracted BufferManager from vvplay_async
