+++ src/vvplay_async_prefetch/mod.rs
@@ -1,3 +1,4 @@
 pub mod network_trace;
 pub mod camera_trace;
 pub mod fetch_request;
+pub mod buffer_manager;
